# internal/controller/http/v1/packet.go
package v1

import (
	"errors"
	"github.com/gofiber/fiber/v2"
	"github.com/hong195/aggregator-sevice/internal/repo"
	"github.com/hong195/aggregator-sevice/internal/usecase/query"
	"net/http"
	"strconv"
)

// @Summary     Find packet
// @Description Find packet by ID
// @ID          history
// @Tags  	    packet
// @Produce     json
// @Success     200 {object} query.DataPacketView
// @Failure     500 {object} response.Error
// @Router      /api/v1/packets/:id [get]
func (r *V1) findPacket(ctx *fiber.Ctx) error {
	idStr := ctx.Params("id")
	packet, err := r.t.Queries.FindDataPacketById.Handle(ctx.Context(), idStr)

	if err != nil {
		r.l.Error(err, "http - v1 - packet")

		if errors.Is(err, repo.ErrInvalidPeriod) {
			return errorResponse(ctx, http.StatusNotFound, "Packet not found")
		}

		return errorResponse(ctx, http.StatusInternalServerError, "database problems")
	}

	return ctx.Status(http.StatusOK).JSON(packet)
}

// @Summary     Get packets by period
// @Description Return packets with max values for the time window [start, end] (Unix ms, UTC).
// @ID          packets-by-period
// @Tags        packets
// @Accept      json
// @Produce     json
// @Param       start query int true  "Start time (Unix ms)"
// @Param       end   query int true  "End time (Unix ms)"
// @Success     200 {array} query.DataPacketView
// @Failure     400 {object} response.Error "missing/invalid query params"
// @Failure     422 {object} response.Error "end < start"
// @Failure     500 {object} response.Error
// @Router      /api/v1/packets/packets [get]
func (r *V1) listPackets(ctx *fiber.Ctx) error {
	startStr := ctx.Query("start")
	endStr := ctx.Query("end")

	if startStr == "" || endStr == "" {
		return errorResponse(ctx, http.StatusBadRequest, "query params 'start' and 'end' are required (unix ms)")
	}

	startMs, err := strconv.ParseInt(startStr, 10, 64)
	if err != nil {
		return errorResponse(ctx, http.StatusBadRequest, "invalid 'start' (must be unix ms)")
	}
	endMs, err := strconv.ParseInt(endStr, 10, 64)
	if err != nil {
		return errorResponse(ctx, http.StatusBadRequest, "invalid 'end' (must be unix ms)")
	}

	q := query.FindDataPacketByPeriodQuery{
		Start: startMs,
		End:   endMs,
	}

	items, handleErr := r.t.Queries.FindDataPacketByPeriod.Handle(ctx.UserContext(), q)
	if handleErr != nil {
		if errors.Is(handleErr, repo.ErrInvalidPeriod) {
			return errorResponse(ctx, http.StatusUnprocessableEntity, "end must be >= start")
		}
		r.l.Error(handleErr, "http - v1 - listPackets")
		return errorResponse(ctx, http.StatusInternalServerError, "storage error")
	}

	return ctx.Status(http.StatusOK).JSON(items)
}